feat: 640 - rename files and path
