+++ dev-data/courses.data.ts
@@ -1,22 +1,22 @@
 import { ROUTES } from '@/core/const';
 import type { Course } from '@/entities/course';
-import angularSecondary from '@/shared/assets/icons/angular-secondary.webp';
-import angular from '@/shared/assets/icons/angular.svg';
+import angular from '@/shared/assets/icons/angular-256.svg';
+import angularSecondary from '@/shared/assets/icons/angular-black-256.svg';
 import awsSecondary from '@/shared/assets/icons/aws-secondary.webp';
 import aws from '@/shared/assets/icons/aws.svg';
-import angularSmall from '@/shared/assets/icons/footer/angular.webp';
+import angularSmall from '@/shared/assets/icons/footer/angular-256.svg';
 import awsDevSmall from '@/shared/assets/icons/footer/aws-dev.webp';
 import awsFundSmall from '@/shared/assets/icons/footer/aws-fundamentals.webp';
-import htmlSmall from '@/shared/assets/icons/footer/html.webp';
-import jsSmall from '@/shared/assets/icons/footer/javascript.webp';
-import nodejsSmall from '@/shared/assets/icons/footer/nodejs.webp';
-import reactSmall from '@/shared/assets/icons/footer/react.webp';
-import javascript from '@/shared/assets/icons/javascript.webp';
-import jsSecondary from '@/shared/assets/icons/js-secondary.webp';
-import nodejs from '@/shared/assets/icons/node.svg';
-import nodejsSecondary from '@/shared/assets/icons/nodejs-secondary.webp';
-import reactSecondary from '@/shared/assets/icons/react-secondary.webp';
-import react from '@/shared/assets/icons/react.svg';
+import htmlSmall from '@/shared/assets/icons/footer/html5-256.svg';
+import jsSmall from '@/shared/assets/icons/footer/javascript-256.svg';
+import nodejsSmall from '@/shared/assets/icons/footer/node-js-256-small.svg';
+import reactSmall from '@/shared/assets/icons/footer/react-256.svg';
+import jsSecondary from '@/shared/assets/icons/javascript-256-black.svg';
+import javascript from '@/shared/assets/icons/javascript-256.svg';
+import nodejs from '@/shared/assets/icons/node-js-256.svg';
+import nodejsSecondary from '@/shared/assets/icons/node-js-black-256.svg';
+import react from '@/shared/assets/icons/react-256.svg';
+import reactSecondary from '@/shared/assets/icons/react-black-256.svg';
 import { COURSE_LINKS, TO_BE_DETERMINED } from '@/shared/constants';
 import { COURSE_TITLES } from 'data';
