+++ src/shared/icons/javascript-icon.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import javascript from '@/shared/assets/icons/js-secondary.webp';
+import javascript from '@/shared/assets/icons/javascript-256-black.svg';
 
 export const JavascriptIcon = () => {
   return <Image src={javascript} alt="JavaScript icon" />;
